Apply fixes for memory leak
As suggested here:
https://github.com/LinusCDE/mad-jetson-ffmpeg/issues/1#issuecomment-1267756761
the following commit is applied to fix a memory leak made by CTCaer here:
https://gitlab.com/switchroot/switch-l4t-multimedia/FFmpeg/-/commit/deac819a

The patch seems to target a different nvv4l2 decoder, but given the similarities,
it seems likely to fix this memory leak as well.

This should resolve #1 (https://github.com/LinusCDE/mad-jetson-ffmpeg/issues/1).